apps/analytics/main.py
from fastapi import FastAPI, HTTPException
from pydantic import BaseModel
import os
from openai import OpenAI
from supabase import create_client, Client
from dotenv import load_dotenv

load_dotenv()

app = FastAPI(title="SmartMenu AI Analytics Agent")

# Initialize clients
supabase: Client = create_client(
    os.getenv("SUPABASE_URL", ""),
    os.getenv("SUPABASE_SERVICE_ROLE_KEY", "")
)
openai_client = OpenAI(api_key=os.getenv("OPENAI_API_KEY"))

class AnalysisRequest(BaseModel):
    tenant_id: str
    action: str
    payload: dict

@app.get("/")
async def root():
    return {"status": "online", "agent": "SmartMenu Analytics"}

@app.post("/v1/reason")
async def reason(request: AnalysisRequest):
    """
    Core AI reasoning endpoint for complex automations.
    """
    try:
        # Example logic: AI-driven menu optimization
        if request.action == "OPTIMIZE_MENU":
            prompt = f"Optimize menu for tenant {request.tenant_id} based on payload: {request.payload}"
            # AI logic goes here
            return {"suggested_changes": [], "reasoning": "Initial agent setup."}
        
        return {"message": "Action not implemented yet."}
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))

if __name__ == "__main__":
    import uvicorn
    uvicorn.run(app, host="0.0.0.0", port=8000)
