apps/analytics/requirements.txt
fastapi
uvicorn
pandas
numpy
supabase
openai
langchain
python-dotenv
requests
